Removed the StrategiesActions.kt class and broke it into different files with the intention of improving the app's structure.
