+++ app/src/main/AndroidManifest.xml
@@ -63,16 +63,13 @@
             android:exported="false"
             android:theme="@style/Theme.AppCompat.NoActionBar" />
         <activity
-            android:name=".StrategiesAndActionsActivity"
+            android:name=".ui.activities.StrategiesActionsActivity"
             android:exported="false"
             android:theme="@style/Theme.AppCompat.NoActionBar" />
         <activity
             android:name=".NotificationsActivity"
             android:exported="true"
             android:theme="@style/Theme.App.NoActionBar" />
-        <activity
-            android:name=".StrategiesActions"
-            android:exported="true" />
         <activity
             android:name=".MainActivity"
             android:exported="true"
