+++ app/src/main/java/com/anxietystressselfmanagement/AwarenessActivity.kt
@@ -6,6 +6,7 @@ import android.os.Bundle
 import android.widget.*
 import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
+import com.anxietystressselfmanagement.ui.activities.StrategiesActionsActivity
 import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.firestore.FirebaseFirestore
 import java.text.SimpleDateFormat
@@ -130,7 +131,7 @@ class AwarenessActivity : AppCompatActivity() {
      * Navigates to the StrategiesAndActionsActivity.
      */
     private fun navigateToStrategies() {
-        val intent = Intent(this, StrategiesActions::class.java)
+        val intent = Intent(this, StrategiesActionsActivity::class.java)
         intent.putExtra("selectedDate", selectedDate)
         startActivity(intent)
         finish()
